LayoutTests/inspector/elements/hide-shortcut-expected.txt

Tests the hide shortcut, which toggles visibility:hidden on the node and it's ancestors. Bug 110641

parent
child


Running: testSetUp

Running: testToggleHideShortcutOn
=== Added hide shortcut ===
=== Parent node is hidden ===
visibility: hidden;
=== Child node is hidden ===
visibility: hidden;

Running: testToggleHideShortcutOff
=== Removed hide shortcut ===
=== Parent node is visible ===
visibility: visible;
=== Child node is visible ===
visibility: visible;

Running: testToggleHideShortcutOnInFrame
=== Added hide shortcut in frame ===
=== Frame node is hidden ===
visibility: hidden;
